move ens model to own file
