+++ models/src/anemoi/models/models/__init__.py
@@ -7,8 +7,8 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-from .encoder_processor_decoder import AnemoiEnsModelEncProcDec
 from .encoder_processor_decoder import AnemoiModelEncProcDec
+from .ens_encoder_processor_decoder import AnemoiEnsModelEncProcDec
 from .hierarchical import AnemoiModelEncProcDecHierarchical
 from .interpolator import AnemoiModelEncProcDecInterpolator
