+++ training/src/anemoi/training/config/model/gnn_ens.yaml
@@ -3,7 +3,7 @@ num_channels: 512
 output_mask: null
 
 model:
-  _target_: anemoi.models.models.encoder_processor_decoder.AnemoiEnsModelEncProcDec
+  _target_: anemoi.models.models.ens_encoder_processor_decoder.AnemoiEnsModelEncProcDec
 
 layer_kernels:
   processor:
