+++ training/src/anemoi/training/config/model/graphtransformer_ens.yaml
@@ -4,7 +4,7 @@ output_mask: null
 cpu_offload: False
 
 model:
-  _target_: anemoi.models.models.encoder_processor_decoder.AnemoiEnsModelEncProcDec
+  _target_: anemoi.models.models.ens_encoder_processor_decoder.AnemoiEnsModelEncProcDec
 
 layer_kernels:
   processor:
